Move store into app, not lib
