+++ src/ambient.d.ts
@@ -7,3 +7,7 @@ declare module 'dotenv';
 declare module 'lodash.pick';
 declare module 'material-ui/*';
 declare module 'react-virtualized';
+
+// NB: This is so that the transpiled version of this file isn't empty,
+// which confuses electron-compile
+export const itsTrue = 'true';
